refactor: rename Chartsv2 to Charts (no v1 exists)
